+++ server/controllers/stars.js
@@ -14,6 +14,7 @@ const getStarsCount = (req, res) => {
 
 const getAverageStarByMovieId = (req, res) => {
   const { movieId } = req.params;
+  if (!movieId) return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   // TODO : movieId validation
   const averageStar = starDao.getAverageStarByMovieId(movieId);
   res.status(200).json(resData.successTrue(resMessage.STAR_GET_SUCCESS, { averageStar }));
@@ -21,6 +22,8 @@ const getAverageStarByMovieId = (req, res) => {
 
 const addStar = (req, res) => {
   const newStar = req.body;
+  if (Object.values(newStar).some(info => !info) || Object.values(newStar).length !== 4)
+    return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   if (starDao.findStarById(newStar.id)) return res.status(400).json(resData.successFalse(resMessage.ID_ALREADY_EXIST));
   if (!userDao.findUserByEmail(newStar.userEmail))
     return res.status(400).json(resData.successFalse(resMessage.EMAIL_NOT_EXIST));
@@ -33,6 +36,7 @@ const addStar = (req, res) => {
 
 const updateStar = (req, res) => {
   const { id, score } = req.body;
+  if (!id || !score) return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   if (!starDao.findStarById(id)) return res.status(400).json(resData.successFalse(resMessage.ID_NOT_EXIST));
   starDao.updateStar(id, score);
   res.status(200).json(resData.successTrue(resMessage.STAR_UPDATE_SUCCESS));
@@ -40,6 +44,7 @@ const updateStar = (req, res) => {
 
 const removeStar = (req, res) => {
   const { id } = req.params;
+  if (!id) return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   if (!starDao.findStarById(id)) return res.status(400).json(resData.successFalse(resMessage.ID_NOT_EXIST));
   starDao.removeStar(id);
   res.status(200).json(resData.successTrue(resMessage.STAR_DELETE_SUCCESS));
@@ -47,12 +52,14 @@ const removeStar = (req, res) => {
 
 const getStarByMovieIdUserEmail = (req, res) => {
   const { movieId, userEmail } = req.params;
+  if (!movieId || !userEmail) return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   const star = starDao.getStarByMovieIdUserEmail(movieId, userEmail);
   res.status(200).json(resData.successTrue(resMessage.STAR_GET_SUCCESS, { star }));
 };
 
 const getStarsByUserEmail = (req, res) => {
   const { userEmail } = req.params;
+  if (!userEmail) return res.status(400).json(resData.successFalse(resMessage.VALUE_INVALID));
   const stars = starDao.getStarsByUserEmail(userEmail);
   res.status(200).json(resData.successTrue(resMessage.STAR_GET_SUCCESS, stars));
 };
